pyproject.toml
[project]
name = "uiautomation"
version = "0.1.0"
description = "iOS Native Apps UI Automation Framework using Appium and Python"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "UIAutomation" }]

dependencies = [
    "appium-python-client>=3.0.0",
    "selenium>=4.15.0",
    "pytest>=7.4.0",
    "pytest-timeout>=2.2.0",
    "pytest-html>=4.1.0",
    "pytest-xdist>=3.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "flake8>=6.1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "settings: Tests for the Settings app",
    "safari: Tests for Safari browser",
    "contacts: Tests for Contacts app",
    "calendar: Tests for Calendar app",
    "photos: Tests for Photos app",
    "smoke: Smoke tests (quick sanity checks)",
    "regression: Full regression tests",
    "slow: Tests that take a long time to run",
    "debug: Debug tests for capturing page source and locator discovery",
    "onboarding: Tests for first-time app launch onboarding screens",
]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = "-v --tb=short"
timeout = 120
